Refactored project to run tests

+++ Sources/Bookstore/Config.swift
@@ -2,7 +2,7 @@
 struct Config {
 
 	static let databaseHost = "localhost"
-	static let databasePort = 5432
+	static let databasePort = Int32(5432)
 	static let userName     = "rfdickerson"
 	static let password     = "password"
 	static let databaseName = "bookstoredb"
